move skull block classes
