+++ src/main/java/xen42/canadamod/CanadaBlocks.java
@@ -47,10 +47,10 @@ import net.minecraft.registry.RegistryKeys;
 import net.minecraft.sound.BlockSoundGroup;
 import net.minecraft.util.Identifier;
 import xen42.canadamod.block.CookingPotBlock;
-import xen42.canadamod.block.MooseSkullBlock;
-import xen42.canadamod.block.MooseWallSkullBlock;
 import xen42.canadamod.block.RubberBlock;
 import xen42.canadamod.block.TreeTapBlock;
+import xen42.canadamod.block.skull.MooseSkullBlock;
+import xen42.canadamod.block.skull.MooseWallSkullBlock;
 import xen42.canadamod.sign.MapleHangingSignBlock;
 import xen42.canadamod.sign.MapleHangingSignBlockEntity;
 import xen42.canadamod.sign.MapleSignBlock;
